refactor: move TABLE_STYLES to constants/tableStyles.ts, update 52 import paths
