+++ frontend/src/pages/inventory/InventorySummaryReport.tsx
@@ -43,7 +43,8 @@ import {
 } from '@mui/icons-material'
 import PageHeader from '@/components/common/PageHeader'
 import { formatCurrency, formatDateTime } from '@/utils/formatters'
-import { TYPOGRAPHY_STYLES, TABLE_STYLES } from '@/constants/typography'
+import { TYPOGRAPHY_STYLES } from '@/constants/typography'
+import { TABLE_STYLES } from '@/constants/tableStyles'
 import { ApiService } from '@/services/api'
 import { useGetEffectivePriceListsQuery } from '@/store/api/priceListApi'
 import { printColors } from '@/styles/printTokens'
