+++ frontend/src/pages/sales/CustomerProfilePage.tsx
@@ -35,7 +35,7 @@ import { useDeleteCustomerMutation } from '@/store/api/salesApi'
 import api from '@/services/api'
 import type { Customer } from '@/types'
 import { CustomerType } from '@/types'
-import { TABLE_STYLES } from '@/constants/typography'
+import { TABLE_STYLES } from '@/constants/tableStyles'
 import { formatCurrency } from '@/utils/currency'
 import { formatDate } from '@/utils/formatters'
 import ConfirmationDialog from '@/components/common/ConfirmationDialog'
