+++ frontend/src/pages/sales/components/OrdersTable.tsx
@@ -11,7 +11,8 @@ import {
   Typography,
 } from '@mui/material'
 
-import { TABLE_STYLES, TYPOGRAPHY_STYLES } from '@/constants/typography'
+import { TABLE_STYLES } from '@/constants/tableStyles'
+import { TYPOGRAPHY_STYLES } from '@/constants/typography'
 import type { SalesOrder } from '@/types'
 
 interface OrderRowProps {
